requirements.txt
pygame-ce==2.5.6
numpy>=2.2,<3
moderngl>=5.12,<6
glcontext>=3.0,<4
trimesh>=4.8,<5
